+++ CHANGELOG.md
@@ -437,3 +437,4 @@ modify it.  Remove query_parms (mutable argument) from calling sequence.  All te
  * **2015-12-24** Replace all query_parm values with one global query_parm in test_vivopump.py.
  * **2015-12-28** Rename `tests` to `test`. PEP8 improvements in utils.py, pubmed.py, pump.py, setup.py, vivo_name.py,
  vivopump.py. Improved comments in `sv.py`
+ * **2015-12-30** Improve education example.